style_transfer/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0
tqdm>=4.50.0
pandas>=1.1.0
transformers>=4.10.0  # Optional, only needed for Vision Transformer implementation
jupyterlab>=3.0.0
jupyter>=1.0.0
ipywidgets>=7.6.0
scikit-image>=0.18.0  # For image processing utilities
opencv-python>=4.5.0  # Additional image processing capabilities
